+++ src/lib/st-select/st-select.component.spec.ts
@@ -8,29 +8,30 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { Component, DebugElement, SimpleChanges, SimpleChange } from '@angular/core';
-import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { FormControl, FormsModule, NgControl, ReactiveFormsModule, Validators } from '@angular/forms';
+import { Component, SimpleChanges, SimpleChange, ViewChild } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormControl, FormsModule, ReactiveFormsModule, Validators, FormGroup } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 
 import { StLabelModule } from '../st-label/st-label.module';
 import { StDropdownMenuModule } from '../st-dropdown-menu/st-dropdown-menu.module';
-import { StSelectComponent } from './st-select.component';
 
 import { StDropDownMenuItem } from '../st-dropdown-menu/st-dropdown-menu.interface';
+import { StSelectComponent } from './st-select.component';
 
 let comboInput: HTMLInputElement;
 let options: StDropDownMenuItem[] = [
-   { label: 'label1', value: 'value1' },
-   { label: 'label2', value: 'value2' },
-   { label: 'label3', value: 'value3' },
-   { label: 'label4', value: 'value4' },
-   { label: 'label5', value: 'value5' },
-   { label: 'label6', value: 'value6' },
-   { label: 'label7', value: 'value7' },
-   { label: 'label8', value: 'value8' },
-   { label: 'label9', value: 'value9' },
-   { label: 'label10', value: 'value10' }
+   {label: 'Select an option', value: undefined},
+   {label: 'label1', value: 'value1'},
+   {label: 'label2', value: 'value2'},
+   {label: 'label3', value: 'value3'},
+   {label: 'label4', value: 'value4'},
+   {label: 'label5', value: 'value5'},
+   {label: 'label6', value: 'value6'},
+   {label: 'label7', value: 'value7'},
+   {label: 'label8', value: 'value8'},
+   {label: 'label9', value: 'value9'},
+   {label: 'label10', value: 'value10'}
 ];
 
 describe('StSelect', () => {
@@ -38,7 +39,6 @@ describe('StSelect', () => {
    let component: StSelectComponent;
    let fixture: ComponentFixture<StSelectComponent>;
    let compiled: any;
-   let de: DebugElement;
 
    beforeEach(async(() => {
       TestBed.configureTestingModule({
@@ -53,7 +53,6 @@ describe('StSelect', () => {
       comboInput = fixture.debugElement.query(By.css('input')).nativeElement;
       component = fixture.componentInstance;
       compiled = fixture.debugElement.nativeElement;
-
    });
 
    it('st-select should have a placeholder', () => {
@@ -63,7 +62,7 @@ describe('StSelect', () => {
       expect(comboInput.placeholder).toEqual(placeholder);
    });
 
-   it('st-select should be init with options', () => {
+   it('st-select should be initialized with options', () => {
       component.options = options;
       fixture.detectChanges();
       expect(component.hasOptions()).toBeTruthy();
@@ -75,7 +74,7 @@ describe('StSelect', () => {
       expect(component.options).toEqual([]);
    });
 
-   it('st-select should change disabled status', () => {
+   it('st-select should change to disabled status', () => {
       fixture.detectChanges();
       expect(component.disabled).toBeFalsy();
 
@@ -88,12 +87,14 @@ describe('StSelect', () => {
       expect(component.disabled).toBeFalsy();
    });
 
-   it('st-select should change value when come from outside', () => {
+   it('st-select should change its value when it comes from outside', () => {
+      component.options = options;
       fixture.detectChanges();
       expect(component.selectedValue).toBeUndefined();
 
-      component.writeValue(options[2]);
+      component.writeValue(options[2].value);
       fixture.detectChanges();
+
       expect(component.selectedValue).toBeDefined();
       expect(component.selectedValue).toEqual(options[2]);
    });
@@ -107,6 +108,7 @@ describe('StSelect', () => {
       component.registerOnChange(fakeFunction);
       component.onChange(options[1]);
       fixture.detectChanges();
+
       expect(fakeFunction).toHaveBeenCalled();
       expect(fakeFunction).toHaveBeenCalledWith(options[1]);
    });
@@ -120,47 +122,39 @@ describe('StSelect', () => {
       component.registerOnTouched(fakeFunction);
       component.onTouched();
       fixture.detectChanges();
+
       expect(fakeFunction).toHaveBeenCalled();
    });
 
-   it('st-select should write value when update forceValidations input', () => {
+   it('st-select should update its value when forceValidations input is true', () => {
       spyOn(component, 'writeValue').and.callThrough();
       fixture.detectChanges();
       expect(component.writeValue).not.toHaveBeenCalled();
 
-      let changeOther: SimpleChanges = { 'options': new SimpleChange([], [], true) };
+      let changeOther: SimpleChanges = {'options': new SimpleChange([], [], true)};
       component.ngOnChanges(changeOther);
       fixture.detectChanges();
       expect(component.writeValue).not.toHaveBeenCalled();
 
-      let change: SimpleChanges = { 'forceValidations': new SimpleChange(false, true, true) };
+      let change: SimpleChanges = {'forceValidations': new SimpleChange(false, true, true)};
       component.ngOnChanges(change);
       fixture.detectChanges();
       expect(component.writeValue).toHaveBeenCalled();
    });
 
-   it('st-select should change option and notify', () => {
+   it('st-select should change the selected option and notify it', () => {
+      let fakeOnChangeFunction: (_: any) => void = jasmine.createSpy('fake');
+      let fakeOnTouchedFunction: () => void = jasmine.createSpy('fake');
+      component.registerOnChange(fakeOnChangeFunction);
+      component.registerOnTouched(fakeOnTouchedFunction);
       component.options = options;
-      fixture.detectChanges();
-
-      let fakeOnChange: (_: any) => void = jasmine.createSpy('fakeOnChange');
-      let fakeOnTouched: () => void = jasmine.createSpy('fakeOnTouched');
-
-      component.registerOnChange(fakeOnChange);
-      component.registerOnTouched(fakeOnTouched);
-      fixture.detectChanges();
-
-      expect(component.selectedValue).toBeUndefined();
-      expect(component.onChange).not.toHaveBeenCalled();
-      expect(component.onTouched).not.toHaveBeenCalled();
-
       component.changeOption(options[2]);
       fixture.detectChanges();
 
       expect(component.selectedValue).toBeDefined();
       expect(component.selectedValue).toEqual(options[2]);
       expect(component.onChange).toHaveBeenCalled();
-      expect(component.onChange).toHaveBeenCalledWith(options[2]);
+      expect(component.onChange).toHaveBeenCalledWith(options[2].value);
       expect(component.onTouched).toHaveBeenCalled();
    });
 
@@ -218,15 +212,6 @@ describe('StSelect', () => {
       expect(component.showError()).toBeFalsy();
    });
 
-   it('st-select should change classes when error', () => {
-      fixture.detectChanges();
-      expect(component.getBarType()).toEqual('st-input-normal-bar sth-input-normal-bar');
-
-      component.errorMessage = 'error';
-      component.forceValidations = true;
-      fixture.detectChanges();
-      expect(component.getBarType()).toEqual('st-input-error-bar sth-input-error-bar');
-   });
 
    it('st-select should focus input', () => {
       component.errorMessage = 'error';
@@ -247,55 +232,32 @@ describe('StSelect', () => {
       let control: FormControl = new FormControl(undefined, [Validators.required]);
       component.selectedValue = options[2];
       component.forceValidations = true;
-      control.setValue(options[2]);
+      control.setValue(options[2].value);
 
       fixture.detectChanges();
 
       component.validate(control);
       expect(component.showError()).toBeFalsy();
 
-      // Third if in checkValidations
       control.setValue(undefined);
-      fixture.detectChanges();
-      component.validate(control);
-      expect(component.showError()).toBeTruthy();
-
-      // Second if in checkValidations
-      control.setValue({ value: 'value' });
-      fixture.detectChanges();
-      component.validate(control);
-      expect(component.showError()).toBeTruthy();
-
-      control.setValue({ label: 'test' });
-      fixture.detectChanges();
-      component.validate(control);
-      expect(component.showError()).toBeTruthy();
-
-      control.setValue('test');
-      fixture.detectChanges();
-      component.validate(control);
-      expect(component.showError()).toBeTruthy();
-
-      // First if in checkValidations
-      control.setValue({});
       fixture.detectChanges();
       component.validate(control);
       expect(component.showError()).toBeTruthy();
    });
 
    it('st-select should validate with custom message', () => {
-      let control: FormControl = new FormControl();
+      let control: FormControl = new FormControl('', [Validators.required]);
       component.selectedValue = options[2];
       component.forceValidations = true;
       component.errorRequiredMessage = 'This field is required';
-      control.setValue(options[2]);
+      control.setValue(options[2].value);
 
       fixture.detectChanges();
       component.validate(control);
       expect(component.showError()).toBeFalsy();
 
-      control.setValue(options[2]);
-      control.setErrors({ 'required': true });
+      control.setValue(options[2].value);
+      control.setErrors({'required': true});
       fixture.detectChanges();
       component.validate(control);
       expect(component.showError()).toBeTruthy();
@@ -305,14 +267,14 @@ describe('StSelect', () => {
       component.validate(control);
       expect(component.showError()).toBeTruthy();
 
-      control.setValue(options[2]);
-      control.setErrors({ 'custom': true });
+      control.setValue(options[2].value);
+      control.setErrors({'custom': true});
       fixture.detectChanges();
       component.validate(control);
       expect(component.showError()).toBeTruthy();
 
       control.setValue(undefined);
-      control.setErrors({ 'required': true });
+      control.setErrors({'required': true});
       fixture.detectChanges();
       component.validate(control);
       expect(component.showError()).toBeTruthy();
@@ -320,7 +282,76 @@ describe('StSelect', () => {
 
 });
 
+describe('Selects in a form', () => {
+   @Component({
+      template: `
+  <form [formGroup]="form" novalidate>
+      <st-select
+            [options]="options"
+            label="Model field" name="modelField"
+            contextualHelp="Select option from list"
+            [errorRequiredMessage]="errorMessage"
+            placeholder="Select option"
+            formControlName="modelField"
+      ></st-select>
+   </form>`
+   })
 
+   class TestStSelectComponent {
+      public form: FormGroup;
+      public options: StDropDownMenuItem[];
+      public control: FormControl = new FormControl(options[1].value, [Validators.required]);
+      public errorMessage: string = 'Error, this field is required';
 
+      @ViewChild(StSelectComponent) select: StSelectComponent;
 
+      constructor() {
+         this.form = new FormGroup({
+            modelField: this.control
+         });
+         this.options = options;
+      }
+   }
 
+   let component: TestStSelectComponent;
+   let fixture: ComponentFixture<TestStSelectComponent>;
+   let select: StSelectComponent;
+
+   beforeEach(async(() => {
+      TestBed.configureTestingModule({
+         imports: [FormsModule, ReactiveFormsModule, StLabelModule, StDropdownMenuModule],
+         declarations: [TestStSelectComponent, StSelectComponent]
+      })
+         .compileComponents();  // compile template and css
+   }));
+
+   beforeEach(() => {
+      fixture = TestBed.createComponent(TestStSelectComponent);
+      component = fixture.componentInstance;
+      select = component.select;
+      fixture.detectChanges();
+   });
+
+   it('when there is a validation error, it should display the error message and change the label and input class', () => {
+      select.changeOption(options[0]);
+      fixture.detectChanges();
+
+      expect(fixture.nativeElement.querySelector('label').classList).toContain('error');
+      expect(fixture.nativeElement.querySelector('.st-input-container').classList).toContain('error');
+      expect(fixture.nativeElement.querySelector('.st-input-error-message').innerHTML).toBe(component.errorMessage);
+   });
+
+   it('if it is disabled, label is also displayed as disabled', () => {
+      component.form.controls.modelField.disable();
+      fixture.detectChanges();
+
+      expect(fixture.nativeElement.querySelector('label').classList).toContain('disabled');
+      expect(fixture.nativeElement.querySelector('.st-input-container').classList).toContain('disabled');
+
+      component.form.controls.modelField.enable();
+      fixture.detectChanges();
+
+      expect(fixture.nativeElement.querySelector('label').classList).not.toContain('disabled');
+      expect(fixture.nativeElement.querySelector('.st-input-container').classList).not.toContain('disabled');
+   });
+});